Finds the two primes that create a semi prime.
Useful for crypto stuff. Definitely not the fastest way to do it but it was fun and it'll find any semi prime given enough time. A couple speed improvements could be made by either prepopulating arrays with prime numbers instead of generating a bunch every time and by making the prime numbers only be generated up to half the semi prime. I have implemented these in another version but the code was not very clean (very large arrays pasted into code) so I left it out of the final version.